+++ pkg/artifactory/datasource/repository/federated/datasource_artifactory_federated_alpine_repository.go
@@ -16,12 +16,12 @@ func DataSourceArtifactoryFederatedAlpineRepository() *schema.Resource {
 
 	alpineFederatedSchema := utilsdk.MergeMaps(
 		local.AlpineLocalSchema,
-		memberSchema,
+		federatedSchema,
 		resource_repository.RepoLayoutRefSchema(rclass, packageType),
 	)
 
 	var packAlpineMembers = func(repo interface{}, d *schema.ResourceData) error {
-		members := repo.(*federated.AlpineFederatedRepositoryParams).Members
+		members := repo.(*federated.AlpineRepositoryParams).Members
 		return federated.PackMembers(members, d)
 	}
 
@@ -36,7 +36,7 @@ func DataSourceArtifactoryFederatedAlpineRepository() *schema.Resource {
 	)
 
 	constructor := func() (interface{}, error) {
-		return &federated.AlpineFederatedRepositoryParams{
+		return &federated.AlpineRepositoryParams{
 			AlpineLocalRepoParams: local.AlpineLocalRepoParams{
 				RepositoryBaseParams: local.RepositoryBaseParams{
 					PackageType: packageType,