+++ pkg/artifactory/resource/repository/federated/federated.go
@@ -7,7 +7,9 @@ import (
 	"net/url"
 	"strings"
 
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/jfrog/terraform-provider-artifactory/v9/pkg/artifactory/resource/repository"
@@ -40,49 +42,57 @@ var PackageTypesLikeGeneric = []string{
 	"vagrant",
 }
 
-type Member struct {
-	Url     string `hcl:"url" json:"url"`
-	Enabled bool   `hcl:"enabled" json:"enabled"`
+type RepoParams struct {
+	Proxy        string `json:"proxy"`
+	DisableProxy bool   `json:"disableProxy"`
 }
 
-var MemberSchemaGenerator = func(isRequired bool) map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"cleanup_on_delete": {
-			Type:        schema.TypeBool,
-			Optional:    true,
-			Default:     false,
-			Description: "Delete all federated members on `terraform destroy` if set to `true`. Caution: it will delete all the repositories in the federation on other Artifactory instances.",
-		},
-		"member": {
-			Type:     schema.TypeSet,
-			Required: isRequired,
-			Optional: !isRequired,
-			Description: "The list of Federated members. If a Federated member receives a request that does not include the repository URL, it will " +
-				"automatically be added with the combination of the configured base URL and `key` field value. " +
-				"Note that each of the federated members will need to have a base URL set. Please follow the [instruction](https://www.jfrog.com/confluence/display/JFROG/Working+with+Federated+Repositories#WorkingwithFederatedRepositories-SettingUpaFederatedRepository)" +
-				" to set up Federated repositories correctly.",
-			Elem: &schema.Resource{
-				Schema: map[string]*schema.Schema{
-					"url": {
-						Type:             schema.TypeString,
-						Required:         true,
-						Description:      "Full URL to ending with the repositoryName",
-						ValidateDiagFunc: validation.ToDiagFunc(validation.IsURLWithHTTPorHTTPS),
-					},
-					"enabled": {
-						Type:     schema.TypeBool,
-						Required: true,
-						Description: "Represents the active state of the federated member. It is supported to " +
-							"change the enabled status of my own member. The config will be updated on the other " +
-							"federated members automatically.",
+type Member struct {
+	Url     string `json:"url"`
+	Enabled bool   `json:"enabled"`
+}
+
+var SchemaGenerator = func(isRequired bool) map[string]*schema.Schema {
+	return utilsdk.MergeMaps(
+		repository.ProxySchema,
+		map[string]*schema.Schema{
+			"cleanup_on_delete": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: "Delete all federated members on `terraform destroy` if set to `true`. Caution: it will delete all the repositories in the federation on other Artifactory instances.",
+			},
+			"member": {
+				Type:     schema.TypeSet,
+				Required: isRequired,
+				Optional: !isRequired,
+				Description: "The list of Federated members. If a Federated member receives a request that does not include the repository URL, it will " +
+					"automatically be added with the combination of the configured base URL and `key` field value. " +
+					"Note that each of the federated members will need to have a base URL set. Please follow the [instruction](https://www.jfrog.com/confluence/display/JFROG/Working+with+Federated+Repositories#WorkingwithFederatedRepositories-SettingUpaFederatedRepository)" +
+					" to set up Federated repositories correctly.",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"url": {
+							Type:             schema.TypeString,
+							Required:         true,
+							Description:      "Full URL to ending with the repositoryName",
+							ValidateDiagFunc: validation.ToDiagFunc(validation.IsURLWithHTTPorHTTPS),
+						},
+						"enabled": {
+							Type:     schema.TypeBool,
+							Required: true,
+							Description: "Represents the active state of the federated member. It is supported to " +
+								"change the enabled status of my own member. The config will be updated on the other " +
+								"federated members automatically.",
+						},
 					},
 				},
 			},
 		},
-	}
+	)
 }
 
-var memberSchema = MemberSchemaGenerator(true)
+var federatedSchema = SchemaGenerator(true)
 
 func unpackMembers(data *schema.ResourceData) []Member {
 	d := &utilsdk.ResourceData{ResourceData: data}
@@ -107,6 +117,15 @@ func unpackMembers(data *schema.ResourceData) []Member {
 	return members
 }
 
+func unpackRepoParams(data *schema.ResourceData) RepoParams {
+	d := &utilsdk.ResourceData{ResourceData: data}
+
+	return RepoParams{
+		Proxy:        d.GetString("proxy", false),
+		DisableProxy: d.GetBool("disable_proxy", false),
+	}
+}
+
 func PackMembers(members []Member, d *schema.ResourceData) error {
 	setValue := utilsdk.MkLens(d)
 
@@ -122,11 +141,68 @@ func PackMembers(members []Member, d *schema.ResourceData) error {
 	}
 
 	errors := setValue("member", federatedMembers)
-	if errors != nil && len(errors) > 0 {
+	if len(errors) > 0 {
 		return fmt.Errorf("failed saving members to state %q", errors)
 	}
 
 	return nil
+}
+
+func configSync(ctx context.Context, repoKey string, m interface{}) diag.Diagnostics {
+	var ds diag.Diagnostics
+
+	tflog.Info(ctx,
+		"triggering synchronization of the federated member configuration",
+		map[string]interface{}{
+			"repoKey": repoKey,
+		},
+	)
+	_, restErr := m.(utilsdk.ProvderMetadata).Client.R().
+		SetPathParam("repositoryKey", repoKey).
+		Post("artifactory/api/federation/configSync/{repositoryKey}")
+	if restErr != nil {
+		ds = append(ds, diag.Diagnostic{
+			Severity: diag.Warning,
+			Summary:  "failed to trigger synchronization of the federated member configuration",
+			Detail:   restErr.Error(),
+		})
+	}
+
+	return ds
+}
+
+func createRepo(unpack unpacker.UnpackFunc, read schema.ReadContextFunc) schema.CreateContextFunc {
+	return func(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+		var ds diag.Diagnostics
+		ds = append(ds, repository.Create(ctx, d, m, unpack)...)
+		if ds.HasError() {
+			return ds
+		}
+
+		ds = append(ds, configSync(ctx, d.Id(), m)...)
+		if ds.HasError() {
+			return ds
+		}
+
+		return append(ds, read(ctx, d, m)...)
+	}
+}
+
+func updateRepo(unpack unpacker.UnpackFunc, read schema.ReadContextFunc) schema.UpdateContextFunc {
+	return func(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+		var ds diag.Diagnostics
+		ds = append(ds, repository.Update(ctx, d, m, unpack)...)
+		if ds.HasError() {
+			return ds
+		}
+
+		ds = append(ds, configSync(ctx, d.Id(), m)...)
+		if ds.HasError() {
+			return ds
+		}
+
+		return append(ds, read(ctx, d, m)...)
+	}
 }
 
 func deleteRepo(_ context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
@@ -173,9 +249,9 @@ func deleteRepo(_ context.Context, d *schema.ResourceData, m interface{}) diag.D
 func mkResourceSchema(skeema map[string]*schema.Schema, packer packer.PackFunc, unpack unpacker.UnpackFunc, constructor repository.Constructor) *schema.Resource {
 	var reader = repository.MkRepoRead(packer, constructor)
 	return &schema.Resource{
-		CreateContext: repository.MkRepoCreate(unpack, reader),
+		CreateContext: createRepo(unpack, reader),
 		ReadContext:   reader,
-		UpdateContext: repository.MkRepoUpdate(unpack, reader),
+		UpdateContext: updateRepo(unpack, reader),
 		DeleteContext: deleteRepo,
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
@@ -192,7 +268,10 @@ func mkResourceSchema(skeema map[string]*schema.Schema, packer packer.PackFunc, 
 				Version: 2,
 			},
 		},
-		CustomizeDiff: repository.ProjectEnvironmentsDiff,
+		CustomizeDiff: customdiff.All(
+			repository.ProjectEnvironmentsDiff,
+			repository.VerifyDisableProxy,
+		),
 	}
 }